+++ react/components/SearchResult.js
@@ -15,14 +15,14 @@ class SearchResult extends Component {
   static propTypes = searchResultPropTypes
 
   static defaultProps = {
-    hiddenFacets: {
-      layoutMode1: LAYOUT_MODE[0].value,
-      layoutMode2: LAYOUT_MODE[1].value,
+    mobileLayout: {
+      mode1: LAYOUT_MODE[0].value,
+      mode2: LAYOUT_MODE[1].value,
     },
   }
 
   state = {
-    galleryLayoutMode: this.props.hiddenFacets.layoutMode1 || LAYOUT_MODE[0].value,
+    mobileLayoutMode: this.props.mobileLayout.mode1,
     showLoadingAsOverlay: false,
     // The definitions bellow are required because
     // on SSR the getDerivedStateFromProps isn't called
@@ -39,15 +39,15 @@ class SearchResult extends Component {
     hiddenFacets: this.props.hiddenFacets,
   }
 
-  handleLayoutChange = e => {
+  handleMobileLayoutChange = e => {
     e.preventDefault()
 
-    const defaultModes = [this.props.hiddenFacets.layoutMode1, this.props.hiddenFacets.layoutMode2]
-    const modeIndex = (defaultModes.indexOf(this.state.galleryLayoutMode) + 1) % 2
-    const currentMode = defaultModes[modeIndex]
+    const modes = [this.props.mobileLayout.mode1, this.props.mobileLayout.mode2]
+    const modeIndex = (modes.indexOf(this.state.mobileLayoutMode) + 1) % 2
+    const currentMode = modes[modeIndex]
 
     this.setState({
-      galleryLayoutMode: currentMode,
+      mobileLayoutMode: currentMode,
     })
   }
 
@@ -56,9 +56,9 @@ class SearchResult extends Component {
     // so we can show the previous products when the
     // overlay is on the screen
     if (!props.loading) {
-      if (!props.hiddenFacets.layoutMode1 && !props.hiddenFacets.layoutMode2) {
-        props.hiddenFacets.layoutMode1 = LAYOUT_MODE[0].value
-        props.hiddenFacets.layoutMode2 = LAYOUT_MODE[1].value
+      if (!props.mobileLayout.mode1 && !props.mobileLayout.mode2) {
+        props.mobileLayout.mode1 = LAYOUT_MODE[0].value
+        props.mobileLayout.mode2 = LAYOUT_MODE[1].value
       }
 
       const {
@@ -121,7 +121,7 @@ class SearchResult extends Component {
       runtime: { hints: { mobile } },
     } = this.props
     const {
-      galleryLayoutMode,
+      mobileLayoutMode,
       recordsFiltered,
       products,
       brands,
@@ -188,7 +188,7 @@ class SearchResult extends Component {
                 products={products}
                 summary={summary}
                 className="bn"
-                layoutMode={galleryLayoutMode}
+                mobileLayoutMode={mobileLayoutMode}
               />
             )}
             {children}
@@ -202,8 +202,8 @@ class SearchResult extends Component {
           {mobile && <div className={`${searchResult.switch} flex justify-center items-center`}>
             <div className="dn-ns db-s">
               <LayoutModeSwitcher
-                activeMode={galleryLayoutMode}
-                onChange={this.handleLayoutChange}
+                activeMode={mobileLayoutMode}
+                onChange={this.handleMobileLayoutChange}
               />
             </div>
           </div>}