+++ react/index.js
@@ -85,27 +85,34 @@ SearchResultQueryLoader.getSchema = props => {
     type: 'object',
     properties: {
       ...querySchema,
-      hiddenFacets: {
-        title: 'editor.search-result.hiddenFacets',
+      mobileLayout: {
+        title: 'editor.search-result.mobileLayout',
         type: 'object',
         isLayout: true,
         properties: {
-          layoutMode1: {
-            title: 'Layout Mode Switcher 1',
+          mode1: {
+            title: 'editor.search-result.mobileLayout.mode1',
             type: 'string',
             default: LAYOUT_MODE[0].value,
             enum: LAYOUT_MODE.map(opt => opt.value),
             enumNames: LAYOUT_MODE.map(opt => opt.label),
             isLayout: true,
           },
-          layoutMode2: {
-            title: 'Layout Mode Switcher 2',
+          mode2: {
+            title: 'editor.search-result.mobileLayout.mode2',
             type: 'string',
             default: LAYOUT_MODE[1].value,
             enum: LAYOUT_MODE.map(opt => opt.value),
             enumNames: LAYOUT_MODE.map(opt => opt.label),
             isLayout: true,
           },
+        },
+      },
+      hiddenFacets: {
+        title: 'editor.search-result.hiddenFacets',
+        type: 'object',
+        isLayout: true,
+        properties: {
           brands: {
             title: 'editor.search-result.hiddenFacets.brands',
             type: 'boolean',